requirements.txt
Flask
Numpy
Pillow